Add countEqClassesByRowSum

+++ libs/algorithms/array/array.c
@@ -16,7 +16,7 @@ void outputArray_(const int *const a, const size_t n) {
     printf("\n");
 }
 
-void append_(int *const a, size_t *const n, const int value) {
+void append_(int *const a, int *const n, const int value) {
     a[*n] = value;
     (*n)++;
 }
